Zmieniono typ parametru metody Szescian z int na double

+++ Github_Kalkulator_Z_Galeziami/Kalkulator.cs
@@ -14,7 +14,7 @@ namespace Github_Kalkulator_Z_Galeziami
             return a * a;
         }
 
-        public static double Szescian(int a)
+        public static double Szescian(double a)
         {
             return a * a * a;
         }
